rename: jdbcTemplate 사용 repository 구분
