+++ src/main/java/gift/repository/jdbcTemplate/WishListJdbcRepository.java
@@ -1,4 +1,4 @@
-package gift.repository;
+package gift.repository.jdbcTemplate;
 
 import gift.domain.Product;
 import gift.dto.WishProduct;
